+++ packages/renderer/src/renderer/screens/Tabs/index.tsx
@@ -16,7 +16,7 @@ import { useDebug } from '@app/hooks/useDebug';
 import { TabsWrapper } from './Wrappers';
 import { Tab } from './Tab';
 import { version } from '../../../../package.json';
-import { ResizeToggles } from './WindowToggles';
+import { ResizeToggles } from './ResizeToggles';
 
 export const Tabs: React.FC = () => {
   useDebug(window.myAPI.getWindowId());
